+++ src/core/analytics/engines/index.ts
@@ -16,7 +16,7 @@ import type { AnalyticEngine } from 'core/analytics/interface';
  * @param [details] - event details
  */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars-experimental
-const sendAnalyticsEvent: AnalyticEngine = (event, details) => {
+const sendAnalyticsEvent: AnalyticEngine = (...args: unknown[]) => {
 	// Need some implementation
 };